@@ -1,15 +1,13 @@
 using System;
-using BaGet.Configuration;
 using BaGet.Core;
-using BaGet.Core.Server.Extensions;
-using BaGet.Extensions;
+using BaGet.Hosting;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.SpaServices.ReactDevelopmentServer;
 using Microsoft.AspNetCore.SpaServices.StaticFiles;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace BaGet
 {
@@ -24,49 +22,38 @@ namespace BaGet
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.ConfigureBaGet(Configuration, httpServices: true);
+            services.ConfigureHttpServices();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
+            var options = Configuration.Get<BaGetOptions>();
+
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
                 app.UseStatusCodePages();
             }
 
-            // Run migrations if necessary.
-            var options = Configuration.Get<BaGetOptions>();
-            if (options.RunMigrationsAtStartup && options.Database.Type != DatabaseType.AzureTable)
-            {
-                using (var scope = app.ApplicationServices.CreateScope())
-                {
-                    scope.ServiceProvider
-                        .GetRequiredService<IContext>()
-                        .Database
-                        .Migrate();
-                }
-            }
-
-            
-            app.UsePathBase(options.PathBase);
-            
             app.UseForwardedHeaders();
+            app.UsePathBase(options.PathBase);
+
             app.UseSpaStaticFiles();
+
+            app.UseRouting();
 
             app.UseCors(ConfigureCorsOptions.CorsPolicy);
             app.UseOperationCancelledMiddleware();
 
-            app.UseMvc(routes =>
+            app.UseEndpoints(endpoints =>
             {
-                routes
-                    .MapServiceIndexRoutes()
-                    .MapPackagePublishRoutes()
-                    .MapSymbolRoutes()
-                    .MapSearchRoutes()
-                    .MapPackageMetadataRoutes()
-                    .MapPackageContentRoutes();
+                endpoints.MapServiceIndexRoutes();
+                endpoints.MapPackagePublishRoutes();
+                endpoints.MapSymbolRoutes();
+                endpoints.MapSearchRoutes();
+                endpoints.MapPackageMetadataRoutes();
+                endpoints.MapPackageContentRoutes();
             });
 
             app.UseSpa(spa =>
