@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations;
 using BaGet.Core;
 
-namespace BaGet.Aws.Configuration
+namespace BaGet.Aws
 {
     public class S3StorageOptions
     {
