@@ -4,11 +4,11 @@ using System.Threading.Tasks;
 using BaGet.Protocol.Models;
 using NuGet.Versioning;
 
-namespace BaGet.Core.Content
+namespace BaGet.Core
 {
     /// <summary>
     /// The Package Content resource, used to download NuGet packages and to fetch other metadata.
-    /// 
+    ///
     /// See: https://docs.microsoft.com/en-us/nuget/api/package-base-address-resource
     /// </summary>
     public interface IPackageContentService
@@ -22,7 +22,7 @@ namespace BaGet.Core.Content
         /// <returns>The package's versions, or null if the package does not exist.</returns>
         Task<PackageVersionsResponse> GetPackageVersionsOrNullAsync(
             string packageId,
-            CancellationToken cancellationToken = default);
+            CancellationToken cancellationToken);
 
         /// <summary>
         /// Download a package, or null if the package does not exist.
@@ -37,7 +37,7 @@ namespace BaGet.Core.Content
         Task<Stream> GetPackageContentStreamOrNullAsync(
             string packageId,
             NuGetVersion packageVersion,
-            CancellationToken cancellationToken = default);
+            CancellationToken cancellationToken);
 
         /// <summary>
         /// Download a package's manifest (nuspec), or null if the package does not exist.
@@ -52,7 +52,7 @@ namespace BaGet.Core.Content
         Task<Stream> GetPackageManifestStreamOrNullAsync(
             string packageId,
             NuGetVersion packageVersion,
-            CancellationToken cancellationToken = default);
+            CancellationToken cancellationToken);
 
         /// <summary>
         /// Download a package's readme, or null if the package or readme does not exist.
@@ -66,6 +66,20 @@ namespace BaGet.Core.Content
         Task<Stream> GetPackageReadmeStreamOrNullAsync(
             string id,
             NuGetVersion version,
-            CancellationToken cancellationToken = default);
+            CancellationToken cancellationToken);
+
+        /// <summary>
+        /// Download a package's icon, or null if the package or icon does not exist.
+        /// </summary>
+        /// <param name="id">The package id.</param>
+        /// <param name="version">The package's version.</param>
+        /// <param name="cancellationToken">A token to cancel the task.</param>
+        /// <returns>
+        /// The package's icon stream, or null if the package or icon does not exist. The stream may not be seekable.
+        /// </returns>
+        Task<Stream> GetPackageIconStreamOrNullAsync(
+            string id,
+            NuGetVersion version,
+            CancellationToken cancellationToken);
     }
 }
