@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations;
 using BaGet.Core;
 
-namespace BaGet.Gcp.Configuration
+namespace BaGet.Gcp
 {
     public class GoogleCloudStorageOptions : StorageOptions
     {
